+++ BioLizardStylePython/src/BioLizardStylePython/__init__.py
@@ -28,5 +28,5 @@ if len(font_names) == 0:
 else:
     lato_localname = font_names.pop()  #pick first one, usually there should only be one.. (?)
 
-
 from .utils import *
+from .plotly_template import *